# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Amar Steps - Swagger DOC ",
        "description": "L6 Swagger OpenApi description",
        "contact": {
            "email": "nshakib.se@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "127.0.0.1:8000",
            "description": "Sandbox OpenApi dynamic host server"
        }
    ],
    "paths": {
...
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Register Users.",
                "description": "Returns user name email and token.",
                "operationId": "register",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "Phone Number",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "User Email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User Password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password_confirmation",
                        "in": "query",
                        "description": "Password Confirmation",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Account has been created successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:auth",
                            "read:auth"
                        ]
                    }
                ]
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Registration"
                ],
                "summary": "Login Users.",
                "description": "Returns user details with token",
                "operationId": "login",
                "parameters": [
                    {
                        "name": "phone",
                        "in": "query",
                        "description": "User Phone",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "User password",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User login successfully.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "oauth2_security_example": [
                            "write:auth",
                            "read:auth"
                        ]
                    }
                ]
            }
        },
        "/api/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "Get auth information",
                "description": "Returns auth data",
                "operationId": "profile",
                "parameters": [
                    {
                        "name": "authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Retrieve auth profile.",
                        "content": {
                            "application/json": {}
                        }
                    }
                }
            }
        },
        "/api/v1/profile/change-password": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Auth password changed.",
                "description": "Returns updated details.",
                "operationId": "change-password",
                "parameters": [
                    {
                        "name": "Authorization",
                        "in": "header",
                        "description": "Bearer token",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "Accept",
                        "in": "query",
                        "description": "application/json",
                        "required": true,
                        "schema": {
                            "example": {
                                "old_password": "123456",
                                "password": "123abc",
                                "password_confirmation": "123abcd"
                            },
                            "type": "object"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Password has been changed successfully.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": true,
                                        "data": {
                                            "id": 1,
                                            "name": "Admin User",
                                            "email": "admin@example.com",
                                            "phone": "+2992233690457",
                                            "role": "Administrator",
                                            "profile": {
                                                "gender": "male",
                                                "dob": "1970-07-11",
                                                "bio": "Duchess, 'chop off her knowledge, as there was no more to do that,' said the Cat, and vanished. Alice was soon submitted to by all three dates on.",
                                                "address": "7749 Dana Trail Suite 868\\nGibsonmouth, HI 32195-3665"
                                            }
                                        },
                                        "message": "Password has been changed successfully."
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "example": {
                                        "success": false,
                                        "message": "Validation failed.",
                                        "errors": {
                                            "password": "Doesn't match."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}